Create README.md
This project informed me about the difficulties and many different uses code can contain while aiming for the same goal. Using HTML, Console, and Flexbox together to create a similar outcome proved how versatile code can really be when it comes to being a designer and choosing the different routes to go down when designing a project. This assignment really challenged my coding skills and I struggled a lot throughout, but I think in the end I tried my best to design this through the many different types of coding. I found the javascript section of this assignment especially difficult, as I am completely new to javascript and tracking all of the different commands and methods used in the website's functionality proved difficult to understand. I still feel like I'm not fully grasping javascript, but for my first time using it, I don't believe I did too bad. I hope to improve these skills for future endeavors in coding and programming to further my UX and UI skills. I also unfortunately missed the first day of this project and it was really hard keeping up when I didn't fully understand the javascript portion of this project.